+++ examples/get_consumption_data.py
@@ -19,9 +19,8 @@ if __name__ == '__main__':
     customer = client.get_user_profile(customer_id)
     # print(customer)
 
-    # Get metering points, also known as "assets". Each asset has an "assetId" which is needed e.g. to
-    # retrieve energy consumption information for a metering point type asset.
-    metering_points = client.get_assets(customer_id)
+    # Household meters are on /assets/meteringpoints (consumptionMeteringPoint).
+    metering_points = client.get_metering_points(customer_id)
     # print(metering_points)
 
     # Get daily usage for the month of January 2023 for the first metering point. Yes, this means TimeSpan.MONTHLY. If
@@ -30,12 +29,16 @@ if __name__ == '__main__':
     january_energy = client.get_energy(customer_id, asset_id, TimeSpan.MONTHLY, 2023, 1, 1)
     # print(january_energy)
 
-    # You're free to do whatever you want with the data, but for the sake of this example, let's filter out the future
-    # days (i.e. the days with zero consumption) and show just the consumption per day.
-    january_energy_filtered = [data for data in january_energy['results'][0]['data'] if data['consumption'] is not None]
+    # Rows may use totalConsumption (current API) or consumption (older API).
+    from pycaruna import energy_kwh
+
+    january_energy_filtered = [
+        data for data in january_energy["results"][0]["data"]
+        if energy_kwh(data) is not None
+    ]
     january_energy_mapped = list(map(lambda data: {
-        'date': data['timestamp'],
-        'consumption': data['consumption'],
+        "date": data["timestamp"],
+        "consumption": energy_kwh(data),
     }, january_energy_filtered))
 
     print(json.dumps(january_energy_mapped, indent=2))