+++ pycaruna/__init__.py
@@ -1,2 +1,16 @@
-from .client import CarunaPlus, TimeSpan
 from .authenticator import Authenticator
+from .client import CarunaPlus, TimeSpan
+from .exceptions import CarunaApiError, CarunaAuthError, CarunaError
+from .utils import customer_ids_from_user, energy_kwh, normalize_energy
+
+__all__ = [
+    "Authenticator",
+    "CarunaPlus",
+    "CarunaApiError",
+    "CarunaAuthError",
+    "CarunaError",
+    "TimeSpan",
+    "customer_ids_from_user",
+    "energy_kwh",
+    "normalize_energy",
+]